requirements.txt
requests==2.31.0
flask==2.3.2
tiktoken==0.4.0
google-generativeai